Adjust LPJ-GUESS individual's biomass

This function adjusts an LPJ-GUESS individual by calling the LPJ-GUESS allocation function (compiled C++) with a given biomass change. It updates the individual biomass pools directly, and also returns, in a list further adjustments to the litter pools.

adjust.biomass.LPJGUESS(
  individual,
  rel.change,
  sla,
  wooddens,
  lifeform,
  k_latosa,
  k_allom2,
  k_allom3
)

Arguments

individual

A nested list which encapsulates an LPJ-GUESS 'Individual' as read from a binary state file

rel.change

A numeric by which to scale the density and C and N pools

sla

The SLA (specific leaf area) (per PFT parameter)

wooddens

Wood density (kgC/m^2) (per PFT parameter)

lifeform

An integer code for the lifeform of this individual (cohort): 1 = Tree, 2 = Grass

The changes in C pools are determined by the allocation. The changes in the N pools are designed to maintain the pre-exisiing C:N ratios, so N is just scaled using the updated C with the initial C:N ratio. The N storage pools (nstore_longterm and nstore_labile) don't have pre-existing C:N ratios, so they are just scaled by the overall biomass change (the 'rel.change' argument to the function).

Note that after this function is called the function allometry should be used to update the individual and to check that the newly updated individual has a 'valid' allometry. The litter pools should also be updated. This is implemented in the updateState function following the call to this adjustBiomass function.

k_latosa

The leaf area to sapwood area ratio (per PFT parameter)

k_allom2, k_allom3,

Allometry coefficients (per PFT parameters)

crownarea_max

Maximum allowed crown area (m^2) (per PFT parameter)

Value

the scaled 'individual' (the initial nested list with update values)

Author

Matthew Forrest