updateState.LPJGUESS

Adjust LPJ-GUESS state variables based on input parameters.

update_state_LPJGUESS(
  model.state,
  dens.initial,
  dens.target,
  cmass.initial,
  cmass.target,
  HEIGHT_MAX = 150
)

Arguments

model.state

A large multiply-nested list containing the entire LPJ-GUESS state as read by function readStateBinary.LPJGUESS

dens.initial

A numeric vector of the initial stand-level stem densities (indiv/m^2) as named numeric vector with one entry per PFT/species, with the names being the PFT/species codes. These values should be produced using state data assimilation from function XXXXXX.

dens.target

A numeric vector of the target stand-level stem densities (indiv/m^2) as named numeric vector with one entry per PFT/species, with the names being the PFT/species codes. These values should be produced using state data assimilation from function XXXXXX

cmass.initial

Initial carbon mass (kgC/m^2).

cmass.target

A numeric vector of the target stand-level biomasses (kgC/m^2) as named numeric vector with one entry per PFT/species, with the names being the PFT/species codes. These values should be produced using state data assimilation from function XXXXXX

HEIGHT_MAX

Maximum allowed height of an individual. This is the maximum height that a tree can have. This is hard-coded in LPJ-GUESS to 150 m, but for SDA that might be unrealistically big, so this argument allows adjustment.

Value

And updated model state (as a big old list o' lists)

Details

Adjust LPJ-GUESS state

Author

Matthew Forrest