Adjust LPJ-GUESS individual's biomass

This function adjusts an LPJ-GUESS individual by calling the LPJ-GUESS allocation function (compiled C++) with a given biomass change. It updates the individual biomass pools directly, and also returns, in a list further adjustments to the litter pools.

Usage

adjust.biomass.LPJGUESS(
  individual,
  biomass.inc,
  sla,
  wooddens,
  lifeform,
  k_latosa,
  k_allom2,
  k_allom3,
  trace = TRUE
)

Arguments

individual

A nested list which encapsulates an LPJ-GUESS 'Individual' as read from a binary state file

biomass.inc

A numeric value specifying the biomass increment to be applied.

sla

The SLA (specific leaf area) (per PFT parameter)

wooddens

Wood density (kgC/m^2) (per PFT parameter)

lifeform

An integer code for the lifeform of this individual (cohort): 1 = Tree, 2 = Grass

k_latosa

The leaf area to sapwood area ratio (per PFT parameter)

k_allom2, k_allom3,

Allometry coefficients (per PFT parameters)

trace

Logical; if TRUE, prints details of biomass adjustment process.

The changes in C pools are determined by the allocation. The changes in the N pools are designed to maintain the pre-exisiing C:N ratios, so N is just scaled using the updated C with the initial C:N ratio. The N storage pools (nstore_longterm and nstore_labile) don't have pre-existing C:N ratios, so they are just scaled by the overall biomass change (the 'rel.change' argument to the function).

Note that after this function is called the function allometry should be used to update the individual and to check that the newly updated individual has a 'valid' allometry. The litter pools should also be updated. This is implemented in the update.state.LPJGUESS function following the call to this adjust.biomass.LPJGUESS function.

Value

the scaled 'individual' (the initial nested list with update values)

Author

Matthew Forrest